Foundational Level Design Tools
Made adjustments to the way customers spawn so that developers can see what the max values are for each time phase. Created a skip time phase tool, but there isn't a rewind. Adjusted code dependencies, but there need to be a lot more changes throughout the project to make it more modular. Blocked by unknown main scene development, and a lack of prefabs.