Final Sprint 2 Adjustments
Added the new sprites to the assets folder, and placed some of them into the scene.

Created a trash so players can throw out ingredients they accidentally picked up.

Began adjusting the ingredient sprites for when they are placed and such, but did not put that in the build.

Stopped the customer view from moving with the cameras (removing how it was in 2D)

Adjusted the customer so that its order shows up at the end of its animation, and only begins its timer at that point.

Adjusted shared area for possibly upright sprite, though this needs to be discussed w/ Veronica next sprint.

Adjusted programming prefabs so that they are all box colliders (non 2D)